requirements.txt
requests
python-telegram-bot
python-telegram-bot-raw